parralel class mixin
